fixed type bug

+++ software-renderer/src/draw.h
@@ -16,7 +16,7 @@ void setPixel(struct TgaImage *image, size_t x, size_t y, uint32_t colour);
 void fill(struct TgaImage *image, uint32_t colour);
 
 //Bresenham's line algorithm
-void line(struct TgaImage *image, size_t x0, int y0, int x1, int y1, uint32_t colour);
+void line(struct TgaImage *image, int x0, int y0, int x1, int y1, uint32_t colour);
 
 //draw loaded obj file into pixel array (ignoring z values)
 bool drawModel(struct TgaImage *image, struct ObjModel *model);
